# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.4.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}/findings:
    get:
      summary: List findings across every scan on this repository
      operationId: listRepositoryFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
        - name: skill
          in: query
          description: Only return findings produced by scans of this skill (e.g. `semgrep`, `security-deep-dive`).
          schema: { type: string }
        - name: scan_group
          in: query
          description: Only return findings from scans in this parallel batch, so an in-flight audit skill reads only its siblings' findings.
          schema: { type: string }
        - name: sub_path
          in: query
          description: Only return findings for this monorepo sub-package (e.g. `activesupport`), so a repository can be exported or bundled one sub-package at a time.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Stream one finding into the concurrent-finding log
      description: >
        Records a single finding emitted mid-scan so sibling scans in the same
        `scan_group` can read it before this scan completes. The body is one
        finding in the same shape it has in the skill's `report.json`; the
        scan's identity (scan id, repository, commit, sub-path) is stamped from
        the bearer token, never trusted from the body. The finding still
        belongs in the final `report.json`, which reconciles against the
        streamed row by fingerprint rather than duplicating it.
      operationId: streamRepositoryFinding
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /findings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: Fetch one finding with its six-step prose
      operationId: getFinding
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update mutable fields on a finding
      operationId: patchFinding
      description: |
        Each field value is written through the history-tracked path.
        Skills write with source=model_suggested; the browser UI (behind
        its own session) writes with source=analyst. Unknown field names
        are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fields]
              properties:
                by:
                  type: string
                  description: Optional author string for the history row.
                fields:
                  type: object
                  additionalProperties: { type: string }
                  description: |
                    Map of field name to new value. Allowed fields:
                    title, severity, status, cwe, location, affected, reachability, quality_tier,
                    cve_id, ghsa_id, cvss_vector, cvss_v4_vector, fix_version, fix_commit,
                    resolution, disclosure_draft, suggested_recipients, assignee,
                    suggested_fix, suggested_fix_commit, breaking_change, breaking_change_rationale,
                    exploited_in_wild, exploited_in_wild_evidence, mitigation, mitigation_semgrep,
                    release_tag, release_url, last_revalidate_verdict.
      responses:
        '204': { description: Updated }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/repositories/{id}/findings:
    get:
      summary: Export findings for a repository (JSONL, or a shareable bundle)
      description: |
        With `format=jsonl` (the default) returns one JSON object per line
        (`application/x-ndjson`), each a `FindingExport` carrying the full set
        of `db.Finding` columns. With `format=bundle` returns a single
        self-contained JSON document (the minimal ingest shape) that
        round-trips through `/v1/import`. The default bundle is share-safe: it
        carries finding substance (the six-step audit narrative, reachability,
        sinks, VID, provenance) but not the analyst's private workspace or the
        recipient's triage — status, assignee, resolution, notes, and
        communications stay out.

        Adding `include=all` (bundle only) promotes the bundle to an archival
        superset for round-tripping a repo's findings back into your own
        instance: on top of the default it carries the enrichment and disclosure
        work product (snippet, affected, fix_version, CVE/GHSA id, CVSS vectors,
        mitigation, breaking-change verdict, dup-check, disclosure draft,
        exploited-in-wild, upstream fix commit) and the notes, communications,
        and references child records. Instance-local lifecycle (status,
        resolution, assignee), the change history, release-watch, and labels are
        never carried. It is opt-in because notes/communications are internal and
        snippet embeds verbatim source, which a cross-party share must not leak.

        Adding `encrypt=1` (bundle only) wraps the bundle in armored age,
        encrypted to every configured recipient. Adding `scope=findings` (bundle
        only) curates the bundle to the Findings bucket (deep-dive and vuln-scan
        audits plus operator imports), dropping per-repo scanner noise. No bearer
        token; localhost-only.
      operationId: exportRepositoryFindings
      security: []
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: format
          in: query
          description: |
            `jsonl` (default) streams NDJSON; `bundle` returns one JSON sharing
            document. Other values get a 400.
          schema: { type: string, enum: [jsonl, bundle] }
        - name: encrypt
          in: query
          description: |
            Any non-empty value wraps a `format=bundle` export in armored age,
            encrypted to the configured recipients. Requires `format=bundle`
            and at least one configured recipient, else 400.
          schema: { type: string }
        - name: scope
          in: query
          description: |
            `findings` curates a `format=bundle` export to the Findings bucket
            (deep-dive and vuln-scan audits plus operator imports), excluding
            per-repo scanner noise. Requires `format=bundle`; any other value,
            or use on the cross-repo `/v1/findings` and `/v1/scans` endpoints,
            gets a 400. Omitted: the bundle carries every finding.
          schema: { type: string, enum: [findings] }
        - name: include
          in: query
          description: |
            `all` promotes a `format=bundle` export to the archival superset —
            the operator's enrichment and disclosure work product plus the
            notes/communications/references child records — for lossless
            round-trip into your own instance. The default bundle omits them so a
            cross-party share stays lean and leaks no internal workspace.
            Requires `format=bundle`; any other value, or use on the cross-repo
            `/v1/findings` and `/v1/scans` endpoints, gets a 400.
          schema: { type: string, enum: [all] }
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
      responses:
        '200':
          description: |
            JSONL stream (`format=jsonl`), a JSON sharing bundle
            (`format=bundle`), or armored age ciphertext
            (`format=bundle&encrypt=1`).
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/FindingExport' }
            application/json:
              schema: { $ref: '#/components/schemas/ImportMinimal' }
            text/plain:
              schema:
                type: string
                description: Armored age ciphertext, starting `-----BEGIN AGE ENCRYPTED FILE-----`.
        '400':
          description: Unsupported `format`, `encrypt`/`scope`/`include` without `format=bundle`, an unknown `scope`/`include` value, or encryption requested with no recipients configured.
        '403':
          description: Host header is not localhost
        '404':
          description: Repository not found
  /v1/findings:
    get:
      summary: Export findings across all repositories as JSONL
      description: |
        Returns one JSON object per line (`application/x-ndjson`). No bearer
        token; localhost-only. Each line is a `FindingExport` carrying the
        full set of `db.Finding` columns.
      operationId: exportFindings
      security: []
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [jsonl] }
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
      responses:
        '200':
          description: JSONL stream of FindingExport objects, one per line.
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/FindingExport' }
        '400':
          description: Unsupported format query value, or `encrypt`/`scope`/`include` requested (only valid on per-repository bundle exports).
        '403':
          description: Host header is not localhost
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token does not grant access to this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Malformed query or request body, or a missing required field
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    FindingInput:
      type: object
      description: >
        One finding as it appears in a skill's report.json, streamed into the
        concurrent-finding log. Only title, severity and location are required;
        the other report.json finding fields (cwe, sinks, dup_check, trace, …)
        are accepted and stored. Scan-owned identity fields are ignored.
      required: [title, severity, location]
      additionalProperties: true
      properties:
        title:    { type: string }
        severity: { type: string, enum: [Critical, High, Medium, Low] }
        location: { type: string }
        cwe:      { type: string }
        sinks:    { type: array, items: { type: string } }
        dup_check: { type: string, description: "One sentence: which siblings under this scan_group you compared against and why this one is distinct." }
    FindingSummary:
      type: object
      required: [id, scan_id, repository_id, severity, location]
      properties:
        id:            { type: integer, format: int64 }
        scan_id:       { type: integer, format: int64 }
        repository_id: { type: integer, format: int64 }
        finding_id:    { type: string, description: Short id within the scan report (F1, F2, ...). }
        sinks:         { type: string, description: Comma-joined sink ids. }
        title:         { type: string }
        severity:      { type: string, enum: [Critical, High, Medium, Low] }
        status:        { type: string }
        cwe:           { type: string }
        location:      { type: string }
        vid:           { type: string, description: "Code identifier (VID-...) hashed from the sink code at the scanned commit; correlates findings across tools. Empty when not computed." }
        affected:      { type: string }
        reachability:  { type: string, enum: [reachable, harness_only, unclear] }
        quality_tier:  { type: string, enum: [high, low] }
        cve_id:        { type: string }
        ghsa_id:       { type: string, description: "GitHub Security Advisory id (GHSA-xxxx-xxxx-xxxx), set once the advisory is published." }
        cvss_vector:    { type: string, description: "CVSS v3.x base vector (3.0 or 3.1)." }
        cvss_score:     { type: number, description: "Base score derived from cvss_vector." }
        cvss_v4_vector: { type: string, description: "CVSS v4.0 base vector. Independent of cvss_vector; both may be populated." }
        cvss_v4_score:  { type: number, description: "Base score derived from cvss_v4_vector." }
        fix_version:   { type: string }
        fix_commit:    { type: string }
        resolution:    { type: string, enum: [fix, migrate, workaround, adopt, wontfix, ""] }
        assignee:      { type: string }
        exploited_in_wild: { type: string, enum: [yes, no, ""], description: "Analyst's call on whether this finding is being exploited in the wild. Empty means unknown. Automation never sets this." }
        exploited_in_wild_evidence: { type: string }
        mitigation:    { type: string, description: "Markdown body of operational mitigation guidance from the mitigate skill: workarounds, detection, optionally a semgrep rule (see mitigation_semgrep)." }
        mitigation_semgrep: { type: string, description: "Optional YAML semgrep rule from the mitigate skill. Empty when no rule was warranted." }
        dup_check:     { type: string, description: "Audit agent's one-sentence rationale for why this finding is distinct from the siblings already filed under the same scan_group. Empty for findings from skills that do not emit it." }
        novelty:       { type: string, enum: [unfixed, fixed, unclear, not_checked, ""], description: "Latest upstream novelty state. A touched file remains unclear until revalidate classifies the staged history." }
        production_viability: { type: string, enum: [VIABLE, NON_VIABLE, SAMPLE_OR_TEST, CONDITIONAL_VIABLE, ""], description: "Latest critic assessment of whether the finding can affect a real release build. Empty before assessment." }
        novelty_checked_commit: { type: string, description: "Repository HEAD compared with the finding's scanned commit for novelty." }
        novelty_checked_at: { type: string, format: date-time, nullable: true }
    FindingDetail:
      allOf:
        - $ref: '#/components/schemas/FindingSummary'
        - type: object
          description: Adds the six-step narrative and the disclosure draft/routing fields.
          properties:
            trace:            { type: string }
            boundary:         { type: string }
            validation:       { type: string }
            prior_art:        { type: string }
            reach:            { type: string }
            rating:           { type: string }
            disclosure_draft: { type: string }
            suggested_recipients: { type: string, description: "File-level owners for the finding's location, from CODEOWNERS or recent non-bot committers." }
            sub_path:         { type: string, description: "Monorepo sub-folder the producing scan was scoped to. The finding's location is relative to it; empty means the repo root." }
            verification:
              allOf:
                - $ref: '#/components/schemas/FindingVerification'
              nullable: true
            attack_path:
              allOf:
                - $ref: '#/components/schemas/FindingAttackPath'
              nullable: true
            remediation:
              allOf:
                - $ref: '#/components/schemas/RemediationAttempt'
              nullable: true
    FindingVerification:
      type: object
      required: [id, scan_id, status, report, created_at]
      properties:
        id:         { type: integer, format: int64 }
        scan_id:    { type: integer, format: int64 }
        status:     { type: string, enum: [confirmed, fixed, inconclusive, deferred, not_attempted] }
        score:      { type: number, format: double, minimum: 0, maximum: 1, nullable: true, description: "Fraction of the five fixed verification criteria that passed. Null for legacy reports and reports that remain internally inconsistent after repair." }
        report:     { $ref: '#/components/schemas/VerificationReport' }
        created_at: { type: string, format: date-time }
    FindingAttackPath:
      type: object
      required: [id, scan_id, production_viability, report, created_at]
      properties:
        id:                   { type: integer, format: int64 }
        scan_id:              { type: integer, format: int64 }
        production_viability: { type: string, enum: [VIABLE, NON_VIABLE, SAMPLE_OR_TEST, CONDITIONAL_VIABLE] }
        report:               { $ref: '#/components/schemas/AttackPathReport' }
        created_at:           { type: string, format: date-time }
    AttackPathReport:
      type: object
      required: [production_viability, source_state, reason, counterevidence, attacker_position, preconditions, impact, likelihood, severity, applied_adjustments, facts_that_would_change_the_result]
      additionalProperties: false
      properties:
        production_viability: { type: string, enum: [VIABLE, NON_VIABLE, SAMPLE_OR_TEST, CONDITIONAL_VIABLE] }
        source_state: { type: string, enum: [PRESENT, MOVED, MISSING, UNKNOWN] }
        reason: { type: string }
        counterevidence: { type: array, items: { type: string } }
        attacker_position: { type: string }
        preconditions: { type: array, items: { type: string } }
        impact: { type: string }
        likelihood: { type: string, enum: [likely, plausible, unlikely, unknown] }
        severity: { type: string, enum: [Critical, High, Medium, Low] }
        applied_adjustments:
          type: array
          description: "Reserved for severity-cap calibration; critic emits an empty array."
          items:
            type: object
            required: [kind, reason, severity_before, severity_after]
            additionalProperties: false
            properties:
              kind: { type: string }
              reason: { type: string }
              severity_before: { type: string, enum: [Critical, High, Medium, Low] }
              severity_after: { type: string, enum: [Critical, High, Medium, Low] }
        facts_that_would_change_the_result: { type: array, items: { type: string } }
    RemediationAttempt:
      type: object
      required: [id, attempt, patch_scan_id, base_commit, status, created_at]
      properties:
        id:            { type: integer, format: int64 }
        attempt:       { type: integer, minimum: 1 }
        patch_scan_id: { type: integer, format: int64 }
        base_commit:   { type: string }
        status:        { type: string, enum: [verification_incomplete, bypassed_patch, verified_secure] }
        created_at:    { type: string, format: date-time }
        validation:
          allOf:
            - $ref: '#/components/schemas/RemediationValidation'
          nullable: true
    RemediationValidation:
      type: object
      required: [scan_id, outcome, valid_variants, benign_control_passed, bypass_input, created_at]
      properties:
        scan_id:               { type: integer, format: int64 }
        outcome:               { type: string, enum: [failed_to_bypass, bypassed_patch, inconclusive] }
        valid_variants:        { type: integer, minimum: 0 }
        benign_control_passed: { type: boolean }
        bypass_input:          { type: string }
        created_at:            { type: string, format: date-time }
    VerificationReport:
      type: object
      description: "Immutable verify output containing an evidenced attack tree, three attempt records and the five fixed grading criteria. Reports created before attack-tree support may omit attack_tree. See skills/verify/schema.json for the current output contract."
      required: [status]
      additionalProperties: true
      properties:
        status: { type: string, enum: [confirmed, fixed, inconclusive, deferred, not_attempted] }
        attack_tree:
          type: object
          description: "Claim-specific attack path with a rooted parent-linked node set, per-node evidence, an overall reachable/blocked/unproven/not_attempted verdict, and concrete blockers."
          additionalProperties: true
          properties:
            goal: { type: string }
            root_id: { type: string }
            verdict: { type: string, enum: [reachable, blocked, unproven, not_attempted] }
            nodes:
              type: array
              items: { type: object, additionalProperties: true }
            blockers:
              type: array
              items: { type: string }
        attempts:
          type: array
          minItems: 3
          maxItems: 3
          items: { type: object, additionalProperties: true }
        criteria: { type: object, additionalProperties: true }
    FindingExport:
      description: |
        Full db.Finding row used by the unauthenticated JSONL export. Adds
        the dedup, denormalised, and timestamp columns that the bearer-only
        FindingDetail leaves out.
      allOf:
        - $ref: '#/components/schemas/FindingDetail'
        - type: object
          properties:
            commit:            { type: string }
            fingerprint:       { type: string }
            last_seen_scan_id: { type: integer, format: int64 }
            last_seen_commit:  { type: string }
            seen_count:        { type: integer }
            created_at:        { type: string, format: date-time }
            updated_at:        { type: string, format: date-time }
    ImportMinimal:
      description: |
        Minimal hand-written ingest shape for findings from a pentest writeup or
        a tool with no SARIF emitter. It is also the shape scrutineer's own
        `format=bundle` export emits and `/v1/import` round-trips: the seven base
        fields, the extended fields every bundle carries, and the archival fields
        an `include=all` bundle adds. Everything past the base seven is optional;
        a hand-written report or a bundle produced before a field existed simply
        omits it.
      type: object
      required: [findings]
      properties:
        repository: { type: string, format: uri }
        commit:     { type: string }
        tool:       { type: string, description: "Producer name; defaults to 'manual'." }
        findings:
          type: array
          items:
            type: object
            properties:
              title:       { type: string }
              description: { type: string }
              severity:    { type: string, enum: [critical, high, medium, low] }
              confidence:  { type: string, enum: [high, medium, low] }
              cwe:         { type: string }
              location:    { type: string, description: "path[:line[:col]]" }
              patch:       { type: string, description: "unified diff" }
              # Extended fields every scrutineer bundle carries (six-step audit
              # prose, provenance, correlation).
              commit:       { type: string, description: "Per-finding revision; falls back to the top-level commit." }
              sub_path:     { type: string }
              locations:    { type: string, description: "Newline-joined file:line hit set." }
              sinks:        { type: string, description: "Comma-joined sink ids." }
              vid:          { type: string }
              reachability: { type: string, enum: [reachable, harness_only, unclear] }
              quality_tier: { type: string, enum: [high, low] }
              boundary:     { type: string }
              validation:   { type: string }
              prior_art:    { type: string }
              reach:        { type: string }
              rating:       { type: string }
              fix_commit:   { type: string, description: "Base revision the patch applies to." }
              # Archival fields, emitted only with include=all: the operator's
              # enrichment, disclosure work product, and child records.
              snippet:                    { type: string }
              affected:                   { type: string, description: "Vulnerable version range." }
              fix_version:                { type: string }
              cve_id:                     { type: string }
              ghsa_id:                    { type: string }
              cvss_vector:                { type: string, description: "v3.x base vector; score recomputed on import." }
              cvss_v4_vector:             { type: string, description: "v4.0 base vector; score recomputed on import." }
              mitigation:                 { type: string }
              mitigation_semgrep:         { type: string }
              breaking_change:            { type: string, enum: [breaking, non_breaking, unknown] }
              breaking_change_rationale:  { type: string }
              dup_check:                  { type: string }
              disclosure_draft:           { type: string }
              suggested_recipients:       { type: string, description: "File-level owners for the finding's location, from CODEOWNERS or recent non-bot committers." }
              exploited_in_wild:          { type: string, enum: [yes, no] }
              exploited_in_wild_evidence: { type: string }
              upstream_fix_commit:        { type: string, description: "Real upstream fix commit (distinct from fix_commit)." }
              notes:
                type: array
                items:
                  type: object
                  properties:
                    body:       { type: string }
                    by:         { type: string }
                    created_at: { type: string, format: date-time }
              communications:
                type: array
                items:
                  type: object
                  properties:
                    channel:      { type: string, enum: [email, ghsa, issue, pr, direct, registry] }
                    direction:    { type: string, enum: [outbound, inbound] }
                    actor:        { type: string }
                    body:         { type: string }
                    offered_help: { type: string, enum: [pr, funding, adoption, none] }
                    at:           { type: string, format: date-time }
                    created_at:   { type: string, format: date-time }
              references:
                type: array
                items:
                  type: object
                  properties:
                    url:        { type: string, format: uri }
                    tags:       { type: string, description: "Comma-joined: issue, pr, cve, ghsa, patch, advisory, discussion, article." }
                    summary:    { type: string }
                    created_at: { type: string, format: date-time }
...